Operations with User entity were added
A client can receive (GET HTTP method) all or a specific (by their id)
user and can save (POST HTTP method) a user in the database as well.

User entity was created. It describes the table in the database.

A few changes in the application.properties file and docker-compose file: when the app runs on
the locale machine, it uses "localhost" url to connect the database, but
if the app runs in a docker container, it uses "database" (name of the
database container) url.