Auth was added
Now a client can send HTTP requests to /api/v1/auth URL. /sign-up is
used to sign up (save they to the databse) a user, which is sent by the client.
/sign-in is used to sign in (check if a received user by the server stay
in the database).

A few dtos were modified (for example UserCreationDto was renamed to
UserSignUpDto). UserSignInDto was created.

User entity was modified (useless fields were deleted). Therefore the
"users" table was also modified in the database.

New exception: SignInDataException that is handled by
GlobalExceptionHandler.

The sql scripts were modified according to the new model.