Validation for UserSignUpDto was added
The validation is used to check if email and password are correct. If
something wrong a message about it will be written into a Map, which is
converted into a JSON object.

MethodArgumentNotValidException is an exception that occurs when a some
UserSignUpDto's fields are incorrect. This exception is handled by
GlobalExceptionHandler.