Handling HttpMessageNotReadableException was added
A HttpMessageNotReadableException invokes when a parser (Jackson) cannot
map a json object into an OOP object. In our case sometimes the parser is not able
to parse LocalDate as a json string into a LocalDate Java object. But
now this exception is handled by GlobalExceptionHandler.

RegexUtil was added. It is an object that is used to find a received
date in an exception message. Then the date is used in a message of a
server response.